feat: AuthToken 발급 로직 분리
